ISPN-5298 HotRod millisecond precision for lifespan/maxidle

+++ client/hotrod-client/src/main/java/org/infinispan/client/hotrod/impl/operations/PutOperation.java
@@ -21,7 +21,7 @@ public class PutOperation extends AbstractKeyValueOperation<byte[]> {
 
    public PutOperation(Codec codec, TransportFactory transportFactory,
                        byte[] key, byte[] cacheName, AtomicInteger topologyId,
-                       Flag[] flags, byte[] value, int lifespan, int maxIdle) {
+                       Flag[] flags, byte[] value, long lifespan, long maxIdle) {
       super(codec, transportFactory, key, cacheName, topologyId, flags, value, lifespan, maxIdle);
    }
